// packages/client/src/entities/PlayerKey.ts
export const enum PlayerKey {
    PREDICTION_APPLIED_CONTROLS = "prediction_applied_controls",
    PREDICTION_LERP_PROGRESS = "prediction_lerp_progress",
    PREDICTION_LERP_START_TIME = "prediction_lerp_start_time",
    PREDICTION_LERP_DURATION = "prediction_lerp_duration",
    PREDICTION_INITIAL_POS = "prediction_initial_pos",
    PREDICTION_PREDICTED_POS = "prediction_predicted_pos",
    PREDICTION_TARGET_POS = "prediction_target_pos",
    PREDICTION_RECONCILED_POS = "prediction_reconciled_pos",
    PREDICTION_ACK_TICK = "prediction_ack_tick",

    ERROR_TIMER = "error_timer",

    WORLD_SESSION = "world_session"
}